+++ src/extensions/immunity/immunity.jl
@@ -6,19 +6,19 @@ function weightedResponseArithmeticMean(pathogen::Pathogen, host::Host, evt::Int
         for response in host.responses
             reac_sum += response.type.reactivityCoefficient(
                 host.sequence,
-                response.imprinted_pathogen.sequence,
-                response.matured_pathogen.sequence,
+                isnothing(response.imprinted_pathogen) ? "" : response.imprinted_pathogen.sequence,
+                isnothing(response.matured_pathogen) ? "" : response.matured_pathogen.sequence,
                 pathogen.sequence
             )
             numerator_sum += response.type.reactivityCoefficient(
                 host.sequence,
-                response.imprinted_pathogen.sequence,
-                response.matured_pathogen.sequence,
+                isnothing(response.imprinted_pathogen) ? "" : response.imprinted_pathogen.sequence,
+                isnothing(response.matured_pathogen) ? "" : response.matured_pathogen.sequence,
                 pathogen.sequence
             ) * response.type.specific_coefficient_functions[evt](
                 host.sequence,
-                response.imprinted_pathogen.sequence,
-                response.matured_pathogen.sequence,
+                isnothing(response.imprinted_pathogen) ? "" : response.imprinted_pathogen.sequence,
+                isnothing(response.matured_pathogen) ? "" : response.matured_pathogen.sequence,
                 pathogen.sequence
             )
         end
@@ -37,19 +37,19 @@ function infectionProbabilityArithmeticMean(pathogen::Pathogen, host::Host)
         for response in host.responses
             reac_sum += response.type.reactivityCoefficient(
                 host.sequence,
-                response.imprinted_pathogen.sequence,
-                response.matured_pathogen.sequence,
+                isnothing(response.imprinted_pathogen) ? "" : response.imprinted_pathogen.sequence,
+                isnothing(response.matured_pathogen) ? "" : response.matured_pathogen.sequence,
                 pathogen.sequence
             )
             numerator_sum += response.type.reactivityCoefficient(
                 host.sequence,
-                response.imprinted_pathogen.sequence,
-                response.matured_pathogen.sequence,
+                isnothing(response.imprinted_pathogen) ? "" : response.imprinted_pathogen.sequence,
+                isnothing(response.matured_pathogen) ? "" : response.matured_pathogen.sequence,
                 pathogen.sequence
             ) * response.type.infectionCoefficient(
                 host.sequence,
-                response.imprinted_pathogen.sequence,
-                response.matured_pathogen.sequence,
+                isnothing(response.imprinted_pathogen) ? "" : response.imprinted_pathogen.sequence,
+                isnothing(response.matured_pathogen) ? "" : response.matured_pathogen.sequence,
                 pathogen.sequence
             )
         end
@@ -66,13 +66,13 @@ function weightedResponseWinnerTakesAll(pathogen::Pathogen, host::Host, evt::Int
         for response in host.responses
             reaction = response.type.reactivityCoefficient(
                 host.sequence,
-                response.imprinted_pathogen.sequence,
-                response.matured_pathogen.sequence,
+                isnothing(response.imprinted_pathogen) ? "" : response.imprinted_pathogen.sequence,
+                isnothing(response.matured_pathogen) ? "" : response.matured_pathogen.sequence,
                 pathogen.sequence
             ) * response.type.specific_coefficient_functions[evt](
                 host.sequence,
-                response.imprinted_pathogen.sequence,
-                response.matured_pathogen.sequence,
+                isnothing(response.imprinted_pathogen) ? "" : response.imprinted_pathogen.sequence,
+                isnothing(response.matured_pathogen) ? "" : response.matured_pathogen.sequence,
                 pathogen.sequence
             )
             if reaction < dominant_reaction
@@ -92,13 +92,13 @@ function infectionProbabilityWinnerTakesAll(pathogen::Pathogen, host::Host)
         for response in host.responses
             reaction = response.type.reactivityCoefficient(
                 host.sequence,
-                response.imprinted_pathogen.sequence,
-                response.matured_pathogen.sequence,
+                isnothing(response.imprinted_pathogen) ? "" : response.imprinted_pathogen.sequence,
+                isnothing(response.matured_pathogen) ? "" : response.matured_pathogen.sequence,
                 pathogen.sequence
             ) * response.type.infectionCoefficient(
                 host.sequence,
-                response.imprinted_pathogen.sequence,
-                response.matured_pathogen.sequence,
+                isnothing(response.imprinted_pathogen) ? "" : response.imprinted_pathogen.sequence,
+                isnothing(response.matured_pathogen) ? "" : response.matured_pathogen.sequence,
                 pathogen.sequence
             )
             if reaction < dominant_reaction
@@ -109,5 +109,19 @@ function infectionProbabilityWinnerTakesAll(pathogen::Pathogen, host::Host)
         return dominant_reaction
     else
         return 1.0
+    end
+end
+
+function deNovoResponse(
+        pathogen::Pathogen, host::Host,
+        existing_responses::Dict{Tuple{String,String,String,String},Response},
+        response_types::Dict{String,ResponseType}; response_type_id::String="Default")
+    if haskey(existing_responses, (host.sequence, pathogen.sequence, "", response_type_id))
+        return [existing_responses[(host.sequence, pathogen.sequence, "", response_type_id)]]
+    else
+        return [newResponse!(
+            pathogen, nothing, host.sequence, existing_responses, response_types[response_type_id],
+            parents=MVector{2,Union{Response,Nothing}}([nothing, nothing])
+        )]
     end
 end